test(remix): Update remix E2E tests to avoid sending to Sentry
Part of https://github.com/getsentry/sentry-javascript/issues/11910
