+++ dev-packages/e2e-tests/test-applications/create-next-app/tests/server-transactions.test.ts
@@ -1,11 +1,6 @@
 import { expect, test } from '@playwright/test';
 import { waitForTransaction } from '@sentry-internal/test-utils';
 
-const authToken = process.env.E2E_TEST_AUTH_TOKEN;
-const sentryTestOrgSlug = process.env.E2E_TEST_SENTRY_ORG_SLUG;
-const sentryTestProject = process.env.E2E_TEST_SENTRY_PROJECT;
-const EVENT_POLLING_TIMEOUT = 90_000;
-
 test('Sends server-side transactions to Sentry', async ({ baseURL }) => {
   const transactionEventPromise = waitForTransaction('create-next-app', transactionEvent => {
     return (
