+++ dev-packages/e2e-tests/test-applications/create-remix-app-express-legacy/tests/server-transactions.test.ts
@@ -1,20 +1,9 @@
 import { expect, test } from '@playwright/test';
-import { waitForError, waitForTransaction } from '@sentry-internal/test-utils';
 import { uuid4 } from '@sentry/utils';
 
-test('Sends a loader error to Sentry', async ({ page }) => {
-  const loaderErrorPromise = waitForError('create-remix-app-express-legacy', errorEvent => {
-    return errorEvent.exception.values[0].value === 'Loader Error';
-  });
+import { waitForTransaction } from '@sentry-internal/test-utils';
 
-  await page.goto('/loader-error');
-
-  const loaderError = await loaderErrorPromise;
-
-  expect(loaderError).toBeDefined();
-});
-
-test('Sends form data with action error to Sentry', async ({ page }) => {
+test('Sends form data with action span to Sentry', async ({ page }) => {
   await page.goto('/action-formdata');
 
   await page.fill('input[name=text]', 'test');
@@ -52,25 +41,16 @@ test('Sends a loader span to Sentry', async ({ page }) => {
   expect(loaderSpan).toBeDefined();
   expect(loaderSpan.op).toBe('function.remix.loader');
 });
-
 test('Propagates trace when ErrorBoundary is triggered', async ({ page }) => {
   // We use this to identify the transactions
   const testTag = uuid4();
 
   const httpServerTransactionPromise = waitForTransaction('create-remix-app-express-legacy', transactionEvent => {
-    return (
-      transactionEvent.type === 'transaction' &&
-      transactionEvent.contexts?.trace?.op === 'http.server' &&
-      transactionEvent.tags?.['sentry_test'] === testTag
-    );
+    return transactionEvent.contexts?.trace?.op === 'http.server' && transactionEvent.tags?.['sentry_test'] === testTag;
   });
 
   const pageLoadTransactionPromise = waitForTransaction('create-remix-app-express-legacy', transactionEvent => {
-    return (
-      transactionEvent.type === 'transaction' &&
-      transactionEvent.contexts?.trace?.op === 'pageload' &&
-      transactionEvent.tags?.['sentry_test'] === testTag
-    );
+    return transactionEvent.contexts?.trace?.op === 'pageload' && transactionEvent.tags?.['sentry_test'] === testTag;
   });
 
   page.goto(`/client-error?tag=${testTag}`);
@@ -104,19 +84,11 @@ test('Sends two linked transactions (server & client) to Sentry', async ({ page 
   const testTag = uuid4();
 
   const httpServerTransactionPromise = waitForTransaction('create-remix-app-express-legacy', transactionEvent => {
-    return (
-      transactionEvent.type === 'transaction' &&
-      transactionEvent.contexts?.trace?.op === 'http.server' &&
-      transactionEvent.tags?.['sentry_test'] === testTag
-    );
+    return transactionEvent.contexts?.trace?.op === 'http.server' && transactionEvent.tags?.['sentry_test'] === testTag;
   });
 
   const pageLoadTransactionPromise = waitForTransaction('create-remix-app-express-legacy', transactionEvent => {
-    return (
-      transactionEvent.type === 'transaction' &&
-      transactionEvent.contexts?.trace?.op === 'pageload' &&
-      transactionEvent.tags?.['sentry_test'] === testTag
-    );
+    return transactionEvent.contexts?.trace?.op === 'pageload' && transactionEvent.tags?.['sentry_test'] === testTag;
   });
 
   page.goto(`/?tag=${testTag}`);
