+++ dev-packages/e2e-tests/test-applications/create-remix-app-express/server.mjs
@@ -48,5 +48,5 @@ app.all(
   }),
 );
 
-const port = process.env.PORT || 3000;
+const port = process.env.PORT || 3030;
 app.listen(port, () => console.log(`Express server listening at http://localhost:${port}`));
