+++ dev-packages/e2e-tests/test-applications/create-remix-app-legacy/app/routes/_index.tsx
@@ -15,8 +15,7 @@ export default function Index() {
         value="Capture Exception"
         id="exception-button"
         onClick={() => {
-          const eventId = Sentry.captureException(new Error('I am an error!'));
-          window.capturedExceptionId = eventId;
+          throw new Error('I am an error!');
         }}
       />
       <Link to="/user/5" id="navigation">
