+++ dev-packages/e2e-tests/test-applications/create-remix-app-v2-legacy/app/entry.client.tsx
@@ -28,21 +28,6 @@ Sentry.init({
   tunnel: 'http://localhost:3031/', // proxy server
 });
 
-Sentry.addEventProcessor(event => {
-  if (
-    event.type === 'transaction' &&
-    (event.contexts?.trace?.op === 'pageload' || event.contexts?.trace?.op === 'navigation')
-  ) {
-    const eventId = event.event_id;
-    if (eventId) {
-      window.recordedTransactions = window.recordedTransactions || [];
-      window.recordedTransactions.push(eventId);
-    }
-  }
-
-  return event;
-});
-
 startTransition(() => {
   hydrateRoot(
     document,
